+++ tests/integration_tests/auto_styles.rs
@@ -7,15 +7,14 @@ fn test_style_stroke_colour() {
     let output = transform_str_default(colour_input).unwrap();
     let expected_style = r#".d-red { stroke: red; }"#;
     assert_contains!(output, expected_style);
-    let expected_style = r#"text.d-red, tspan.d-red, text.d-red * { fill: red; stroke: white; }"#;
+    let expected_style = r#"text.d-red, tspan.d-red { fill: red; stroke: white; }"#;
     assert_contains!(output, expected_style);
 
     let colour_input = r#"<svg><rect xy="0" wh="20" class="d-black" text="T"/></svg>"#;
     let output = transform_str_default(colour_input).unwrap();
     let expected_style = r#".d-black { stroke: black; }"#;
     assert_contains!(output, expected_style);
-    let expected_style =
-        r#"text.d-black, tspan.d-black, text.d-black * { fill: black; stroke: white; }"#;
+    let expected_style = r#"text.d-black, tspan.d-black { fill: black; stroke: white; }"#;
     assert_contains!(output, expected_style);
 
     // Check special case that d-none does not set text fill
@@ -23,7 +22,7 @@ fn test_style_stroke_colour() {
     let output = transform_str_default(colour_input).unwrap();
     let expected_style = r#".d-none { stroke: none; }"#;
     assert_contains!(output, expected_style);
-    let unexpected_style = r#"text.d-none, tspan.d-none, text.d-none * { fill: none;"#;
+    let unexpected_style = r#"text.d-none, tspan.d-none { fill: none;"#;
     assert_not_contains!(output, unexpected_style);
 }
 
@@ -33,15 +32,15 @@ fn test_style_fill_colour() {
     let output = transform_str_default(colour_input).unwrap();
     let expected_style = r#".d-fill-red { fill: red; }"#;
     assert_contains!(output, expected_style);
-    let expected_style =
-        r#"text.d-fill-red, tspan.d-fill-red, text.d-fill-red * { fill: white; stroke: black; }"#;
+    let expected_style = r#"text.d-fill-red, tspan.d-fill-red { fill: white; stroke: black; }"#;
     assert_contains!(output, expected_style);
 
     let colour_input = r#"<svg><rect xy="0" wh="20" class="d-fill-lightgrey" text="test"/></svg>"#;
     let output = transform_str_default(colour_input).unwrap();
     let expected_style = r#".d-fill-lightgrey { fill: lightgrey; }"#;
     assert_contains!(output, expected_style);
-    let expected_style = r#"text.d-fill-lightgrey, tspan.d-fill-lightgrey, text.d-fill-lightgrey * { fill: black; stroke: white; }"#;
+    let expected_style =
+        r#"text.d-fill-lightgrey, tspan.d-fill-lightgrey { fill: black; stroke: white; }"#;
     assert_contains!(output, expected_style);
 }
 
@@ -49,13 +48,12 @@ fn test_style_fill_colour() {
 fn test_style_text_colour() {
     let colour_input = r#"<svg><text xy="0" class="d-text-red">Hello!</text></svg>"#;
     let output = transform_str_default(colour_input).unwrap();
-    let expected_style =
-        r#"text.d-text-red, tspan.d-text-red, text.d-text-red * { fill: red; stroke: white; }"#;
+    let expected_style = r#"text.d-text-red, tspan.d-text-red { fill: red; stroke: white; }"#;
     assert_contains!(output, expected_style);
 
     let colour_input = r#"<svg><text xy="0" class="d-text-black">Hello!</text></svg>"#;
     let output = transform_str_default(colour_input).unwrap();
-    let expected_style = r#"text.d-text-black, tspan.d-text-black, text.d-text-black * { fill: black; stroke: white; }"#;
+    let expected_style = r#"text.d-text-black, tspan.d-text-black { fill: black; stroke: white; }"#;
     assert_contains!(output, expected_style);
 }
 
@@ -63,19 +61,18 @@ fn test_style_text_colour() {
 fn test_style_text_attributes() {
     let input = r#"<svg><text xy="0" class="d-text-bold">Hello!</text></svg>"#;
     let output = transform_str_default(input).unwrap();
-    let expected_style =
-        r#"text.d-text-bold, tspan.d-text-bold, text.d-text-bold * { font-weight: bold; }"#;
+    let expected_style = r#"text.d-text-bold, tspan.d-text-bold { font-weight: bold; }"#;
     assert_contains!(output, expected_style);
 
     let input = r#"<svg><text xy="0" class="d-text-italic">Hello!</text></svg>"#;
     let output = transform_str_default(input).unwrap();
-    let expected_style =
-        r#"text.d-text-italic, tspan.d-text-italic, text.d-text-italic * { font-style: italic; }"#;
+    let expected_style = r#"text.d-text-italic, tspan.d-text-italic { font-style: italic; }"#;
     assert_contains!(output, expected_style);
 
     let input = r#"<svg><text xy="0" class="d-text-monospace">Hello!</text></svg>"#;
     let output = transform_str_default(input).unwrap();
-    let expected_style = r#"text.d-text-monospace, tspan.d-text-monospace, text.d-text-monospace * { font-family: monospace; }"#;
+    let expected_style =
+        r#"text.d-text-monospace, tspan.d-text-monospace { font-family: monospace; }"#;
     assert_contains!(output, expected_style);
 
     let text_sizes = vec![
@@ -90,8 +87,7 @@ fn test_style_text_attributes() {
     for (class, size) in text_sizes {
         let input = format!(r#"<svg><text xy="0" class="{class}">Hello!</text></svg>"#);
         let output = transform_str_default(&input).unwrap();
-        let expected_style =
-            format!("text.{class}, tspan.{class}, text.{class} * {{ font-size: {size}px; }}");
+        let expected_style = format!("text.{class}, tspan.{class} {{ font-size: {size}px; }}");
         assert_contains!(output, &expected_style);
     }
 }
@@ -100,18 +96,18 @@ fn test_style_text_attributes() {
 fn test_style_text_outline() {
     let input = r#"<svg><text xy="0" class="d-text-ol">Hello!</text></svg>"#;
     let output = transform_str_default(input).unwrap();
-    let expected_style =
-        r#"text.d-text-ol, tspan.d-text-ol, text.d-text-ol * { stroke-width: 0.5; }"#;
+    let expected_style = r#"text.d-text-ol, tspan.d-text-ol { stroke-width: 0.5; }"#;
     assert_contains!(output, expected_style);
 
     let input = r#"<svg><text xy="0" class="d-text-ol-red">Hello!</text></svg>"#;
     let output = transform_str_default(input).unwrap();
-    let expected_style = r#"text.d-text-ol-red, tspan.d-text-ol-red, text.d-text-ol-red * { stroke: red; stroke-width: 0.5; }"#;
+    let expected_style =
+        r#"text.d-text-ol-red, tspan.d-text-ol-red { stroke: red; stroke-width: 0.5; }"#;
     assert_contains!(output, expected_style);
 
     let input = r#"<svg><text xy="0" class="d-text-ol-thicker">Hello!</text></svg>"#;
     let output = transform_str_default(input).unwrap();
-    let expected_style = r#"text.d-text-ol-thicker, tspan.d-text-ol-thicker, text.d-text-ol-thicker * { stroke-width: 2; }"#;
+    let expected_style = r#"text.d-text-ol-thicker, tspan.d-text-ol-thicker { stroke-width: 2; }"#;
     assert_contains!(output, expected_style);
 }